Revert "Movie 9 implement a list of favorite movies"
