+++ app/src/main/java/com/example/appmovie/Film.kt
@@ -1,11 +1,10 @@
 package com.example.appmovie
 
-data class Film(
+class Film(
     val id: Int,
     val cover: String,
-    val rating: String,
     val header: String,
     val genre: String,
-    val data: Int,
     val time: String
-)
+) {
+}
