+++ app/src/main/java/com/example/appmovie/FilmItem.kt
@@ -1,0 +1,17 @@
+package com.example.appmovie
+
+import androidx.recyclerview.widget.RecyclerView
+import com.example.appmovie.databinding.ItemFilmBinding
+
+class FilmItem(
+    private val binding: ItemFilmBinding
+) : RecyclerView.ViewHolder(binding.root) {
+
+    fun onBind(film: Film) {
+        with(binding) {
+            tvHeader.text = film.header
+            tvGenre.text = film.genre
+            tvTime.text = film.time
+        }
+    }
+}
