+++ app/src/main/java/com/example/appmovie/FilmRepository.kt
@@ -1,0 +1,28 @@
+package com.example.appmovie
+
+object FilmRepository {
+
+    val films = arrayListOf(
+        Film(
+            cover = "@color/main_color_background_app",
+            id = 0,
+            header = "Spiderman",
+            genre = "Action",
+            time = "139 minutes"
+        ),
+        Film(
+            cover = "@color/main_color_background_app",
+            id = 1,
+            header = "Spiderman",
+            genre = "Action",
+            time = "139 minutes"
+        ),
+        Film(
+            cover = "@color/main_color_background_app",
+            id = 2,
+            header = "Spiderman",
+            genre = "Action",
+            time = "139 minutes"
+        )
+    )
+}
