--- app/src/main/java/com/example/appmovie/MyDiffCallback.kt
@@ -1,14 +1,0 @@
-package com.example.appmovie
-
-import androidx.recyclerview.widget.DiffUtil
-
-class MyDiffCallback : DiffUtil.ItemCallback<Film>() {
-
-    override fun areItemsTheSame(oldItem: Film, newItem: Film): Boolean {
-        return oldItem == newItem
-    }
-
-    override fun areContentsTheSame(oldItem: Film, newItem: Film): Boolean {
-        return oldItem == newItem
-    }
-}
