--- app/src/main/java/com/example/appmovie/movie/data/repository/repository/FilmRepository.kt
@@ -1,35 +1,0 @@
-package com.example.appmovie.movie.data.repository.repository
-
-import com.example.appmovie.Film
-
-object FilmRepository {
-    var films = arrayListOf(
-        Film(
-            cover = "https://avatars.mds.yandex.net/get-kinopoisk-image/1704946/6ac42bf1-53e1-46c0-856f-31ab8373b974/1920x",
-            id = 0,
-            header = "Spiderman",
-            genre = "Action",
-            time = "139 minutes",
-            rating = "9.5",
-            data = 2019
-        ),
-        Film(
-            cover = "https://upload.wikimedia.org/wikipedia/ru/6/6e/Spider-Man_—_No_Way_Home_poster.jpg",
-            id = 1,
-            header = "Spiderman1",
-            genre = "Action",
-            time = "139 minutes",
-            rating = "9.5",
-            data = 2019
-        ),
-        Film(
-            cover = "https://avatars.mds.yandex.net/get-kinopoisk-image/1773646/458a7588-35f6-4928-bffa-e38704e75e26/1920x",
-            id = 2,
-            header = "Spiderman2",
-            genre = "Action",
-            time = "139 minutes",
-            rating = "9.5",
-            data = 2019
-        )
-    )
-}
