+++ app/src/main/res/layout/fragment_favorite.xml
@@ -11,10 +11,11 @@
         android:id="@+id/rv_film_favorite"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        android:layout_marginStart="32dp"
-        android:layout_marginTop="140dp"
-        android:layout_marginEnd="32dp"
-        android:orientation="vertical"
         app:layout_constraintBottom_toBottomOf="parent"
-        tools:itemCount="3" />
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent"
+        app:layoutManager="androidx.recyclerview.widget.LinearLayoutManager"
+        android:orientation="vertical"
+        tools:itemCount="3"/>
 </androidx.constraintlayout.widget.ConstraintLayout>
